backend/requirements.txt
Django>=5.0
djangorestframework
openai
python-dotenv
psycopg2-binary
django-cors-headers

